requirements.txt
atproto==0.0.61
python-dotenv==1.1.0
anthropic==0.49.0
pydantic==2.11.3
config==0.5.1